--- internal/tadpoles/helpers.go
@@ -1,22 +1,0 @@
-package tadpoles
-
-import (
-	"github.com/leocov-dev/tadpoles-backup/internal/api"
-	"github.com/leocov-dev/tadpoles-backup/internal/schemas"
-	log "github.com/sirupsen/logrus"
-)
-
-func eventsToAttachments(events []*api.Event) (attachments []*schemas.FileAttachment) {
-	for _, event := range events {
-		for _, eventAttachment := range event.Attachments {
-			// skip pdf files
-			if eventAttachment.MimeType == "application/pdf" {
-				log.Debugf("skipping pdf: %s@%s \n", event.ChildName, event.EventTime)
-				continue
-			}
-			att := schemas.NewFileAttachment(event, eventAttachment)
-			attachments = append(attachments, att)
-		}
-	}
-	return attachments
-}
